requirements.txt
Django==4.1.3
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
drf-spectacular-sidecar>=2022.11.1
dependencies==7.7.0
environ==1.0